throw chart into node as well Need to fix bug
